feat: :sparkles: add new endpoints for seat management and update event endpoints

+++ src/main/java/com/gerardo/swiftentrybackend/domain/Locality/services/LocalityServiceImpl.java
@@ -1,6 +1,6 @@
 package com.gerardo.swiftentrybackend.domain.Locality.services;
 
-import com.gerardo.swiftentrybackend.domain.Event.EventModel;
+import com.gerardo.swiftentrybackend.common.exceptions.ForbiddenOperationException;
 import com.gerardo.swiftentrybackend.domain.Event.repositories.EventRepository;
 import com.gerardo.swiftentrybackend.domain.Locality.LocalityModel;
 import com.gerardo.swiftentrybackend.domain.Locality.dto.request.LocalityRequestDTO;
@@ -8,10 +8,12 @@ import com.gerardo.swiftentrybackend.domain.Locality.dto.request.LocalityUpdateD
 import com.gerardo.swiftentrybackend.domain.Locality.dto.response.LocalityResponseDTO;
 import com.gerardo.swiftentrybackend.domain.Locality.repositories.LocalityRepository;
 import com.gerardo.swiftentrybackend.domain.Locality.utils.LocalityMapper;
-import com.gerardo.swiftentrybackend.common.exceptions.ResourceConflictException;
+import com.gerardo.swiftentrybackend.domain.Reservation.repositories.ReservationSeatRepository;
+import com.gerardo.swiftentrybackend.domain.Seat.repositories.LocalitySeatRepository;
 import com.gerardo.swiftentrybackend.common.exceptions.ResourceNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
@@ -22,25 +24,11 @@ public class LocalityServiceImpl implements LocalityService {
     private final LocalityRepository localityRepository;
     private final EventRepository eventRepository;
     private final LocalityMapper localityMapper;
+    private final LocalitySeatRepository localitySeatRepository;
+    private final ReservationSeatRepository reservationSeatRepository;
 
     @Override
     public LocalityResponseDTO createLocality(LocalityRequestDTO request) {
-        /*EventModel event = eventRepository.findById(request.getEventId())
-                .orElseThrow(
-                        () -> new ResourceNotFoundException("Event with id " + request.getEventId() + " not found"));
-
-        if (localityRepository.existsByNameAndEvent_Id(request.getName(), request.getEventId())) {
-            throw new ResourceConflictException("Locality '" + request.getName() + "' already exists for this event");
-        }
-
-        LocalityModel newLocality = LocalityModel.builder()
-                .event(event)
-                .name(request.getName())
-                .description(request.getDescription())
-                .price(request.getPrice())
-                .capacity(request.getCapacity())
-                .build();*/
-
         return null;
     }
 
@@ -80,10 +68,16 @@ public class LocalityServiceImpl implements LocalityService {
     }
 
     @Override
+    @Transactional
     public void deleteLocality(Long id) {
         if (!localityRepository.existsById(id)) {
             throw new ResourceNotFoundException("Locality with id " + id + " not found");
         }
+        if (reservationSeatRepository.existsByLocalitySeat_Locality_Id(id)) {
+            throw new ForbiddenOperationException(
+                    "Cannot delete locality with id " + id + " because it has existing reservations.");
+        }
+        localitySeatRepository.deleteAllByLocality_Id(id);
         localityRepository.deleteById(id);
     }
 }
